+++ Tests/InstantiatorTest.php
@@ -12,11 +12,15 @@
 namespace Symfony\Component\VarExporter\Tests;
 
 use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\Attributes\IgnoreDeprecations;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\VarExporter\Exception\ClassNotFoundException;
 use Symfony\Component\VarExporter\Exception\NotInstantiableTypeException;
 use Symfony\Component\VarExporter\Instantiator;
 
+#[Group('legacy')]
+#[IgnoreDeprecations]
 class InstantiatorTest extends TestCase
 {
     public function testNotFoundClass()